Moves FormFieldColorPicker
