+++ src/components/Field/FieldTypes.tsx
@@ -2,7 +2,7 @@ import { FieldDefAddress } from "@root/components/Field/FormFieldAddress";
 import { FieldDefAdvancedSelection } from "@root/components/Field/FormFieldAdvancedSelection";
 import { FieldDefCheckbox } from "@root/components/Field/FormFieldCheckbox";
 import { FieldDefChip } from "@root/components/Field/FormFieldChipSingleSelect";
-import { FieldDefColor } from "@root/forms/FormFieldColorPicker/ColorPickerTypes";
+import { FieldDefColor } from "@root/components/Field/FormFieldColorPicker/ColorPickerTypes";
 import { FieldDefDate } from "@root/forms/FormFieldDate/DateField";
 import { FieldDefDropdown } from "@root/forms/FormFieldDropdownSingleSelection";
 import { FieldDefImageUpload } from "@root/forms/FormFieldImageUpload";
