+++ src/components/Form/Col/Col.tsx
@@ -12,7 +12,7 @@ import FormFieldRaw from "@root/forms/FormFieldRaw";
 import FormFieldToggleSwitch from "@root/forms/FormFieldToggleSwitch";
 import { FieldDef } from "@root/components/Field";
 import FormFieldImageVideoLinkDocumentBrowsing from "@root/forms/FormFieldImageVideoLinkDocumentBrowsing";
-import FormFieldColorPicker from "@root/forms/FormFieldColorPicker";
+import FormFieldColorPicker from "@root/components/Field/FormFieldColorPicker";
 import FormFieldDate from "@root/forms/FormFieldDate/DateField";
 import FormFieldAddress from "@root/components/Field/FormFieldAddress";
 import FormFieldTable from "@root/forms/FormFieldTable";
